FastAPI/auth-mongo/backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose[cryptography]
passlib[bcrypt]